refactor: clean up gsheet express route
